S2 width cut
Diffusion parameters update according to Fig 7 and Fig8 in the note below.
1) 304 -> 350
2) 3.6395 -> 3.9656 (To avoid detail study on diffusion constant and electron drift velocity in limit time, just update this parameter here )
After tuning the parameters, the acceptance inside 99% ER band increased a bit, while the rejection power decreased a bit for the outlier events (Table0 and 1). This decreasing may can be compensated by new S1 pattern cut. Suggested to update to the new parameters. 
https://xecluster.lngs.infn.it/dokuwiki/doku.php?id=xenon:xenon1t:yuehuan:analysis:0sciencerun_s2width_update0#comparison_with_diffusion_model_cut_by_jelle_pax_v642