Update sciencerun1.py with new version of MV cut
The Muon Veto cut removes events in coincidence with Muon Veto triggers.  It checks the distance in time (ns) between a reference position inside the waveform and the nearest MV trigger. The event is excluded if the nearest MV trigger falls in a [-2ms,+3ms] time window with respect to the reference position. It requires Proximity minitrees.

https://xe1t-wiki.lngs.infn.it/doku.php?id=xenon:xenon1t:analysis:mv_cut_sr1