Muon Veto cut added (#55)
* Muon Veto cut added

The Muon Veto cut has been added to the list of cuts in sciencerun0.py.

* Update sciencerun0.py

* Update sciencerun0.py

* Update class MuonVeto to StringLichen

* Add MuonVeto to AllEnergy list

* Ignore MuonVeto cut for MC in laxer

* Add MuonVeto placeholder to sciencerun1

(still to be tuned)

* lint style fixes

* Update sciencerun1.py with new version of MV cut

The Muon Veto cut removes events in coincidence with Muon Veto triggers.  It checks the distance in time (ns) between a reference position inside the waveform and the nearest MV trigger. The event is excluded if the nearest MV trigger falls in a [-2ms,+3ms] time window with respect to the reference position. It requires Proximity minitrees.

https://xe1t-wiki.lngs.infn.it/doku.php?id=xenon:xenon1t:analysis:mv_cut_sr1

* Update sciencerun1.py

MV cut is inherited from sciencerun0.py

* New definition of MV cut extended to SR0

* Remove trailing whitespaces