Add cS2 AFT cut (#142)
* Add CS2 area fraction top cut

Adds a cut on cS2 area fraction top. This has a better resolution than S2 AFT (uncorrected) since it accounts for PMTs which are switched off or set to zero gain in pax in the top array. The two new cuts added are CS2AreaFractionTop, which is a 99% acceptance cut setting an upper and lower bound on the cS2 AFT, and CS2AreaFractionTop96p which is a 96% upper-bound-only cut.

For now these are not added to the default cuts lists so you must apply them manually.

For more details see the note at https://xe1t-wiki.lngs.infn.it/doku.php?id=xenon:xenon1t:adam:s2aft:sr1_cs2_cut#rejection

* Move CS2AreaFractionTop to sciencerun0